+++ tests/sandbox_tests/test_concurrent.py
@@ -34,14 +34,14 @@ class ConcurrentTransactionsTestCase(SandboxedNodeAutoBakeTestCase):
 
     def test_2_batch_multiple_calls(self) -> None:
         contract = self.get_contract()
-        txs = [contract.increment(i) for i in range(10)]
+        txs = [contract.increment(i) for i in range(100)]
         opg = self.client.bulk(*txs).autofill()
         opg.sign().inject(
             time_between_blocks=self.TIME_BETWEEN_BLOCKS,
             min_confirmations=1,
             block_timeout=5,
         )
-        self.assertEqual(45, int(contract.storage()))
+        self.assertEqual(4950, int(contract.storage()))
 
     def test_3_send_multiple_calls(self) -> None:
         contract = self.get_contract()